Fix navigation after user dismiss an alarm
Set task affinity of AlarmActivity to empty.

All activities in the same app have the same task affinity by default.
Calling finish() on a foreground activity will bring another background
activity if there is any. Set task affinity to empty, so unrelated
activities will not be in the same task.

As a result, the background MainActivity no longer comes to foreground
when user dismiss an alarm.