Fix behaviour of snoozed alarms when they are missed
Previously, when the user dismiss all missed alarm, the snooze counter
is not set back to zero. As a result, they keep being displayed as a
snoozed alarm although they are not.

This issue is solved by setting the snooze counter of the alarms to zero
when they are dismissed.