Fix weekly repeat alarm calculation
Before the calculation start, the result date is set to the first day of
the week which the start date is in. When rolling through the first
week, the program only checks if that time is after now, but not if that
time is after the start date. As a result, alarm with a start date in
the future might go off before the start date.

Add a check to ensure the next time is after the start date.
Then recalculate the next time of not snoozed alarms to correct any
error due to previous incorrect calculation.