Add database backup/restore function
Allow to backup/restore the database for testing.

Backup/restore is achieved by copying the database file from/to external
storage. This function does not come with logic for validating the
database, therefore, it should ONLY be used for testing, NOT for
production. It should be hidden on release versions.