Implement foreground service for firing alarm
The foreground service is used to make sure the alarm goes off after
being cleared from recent apps screen in API level 26 and above.

In order for the foreground service to work, a notification indicating
the firing alarm is also implemented, which allows user to dismiss an
alarm in a more convenient manner.

Firing multiple alarms at the same time is also implemented. The alarms
will be in a queue and will be processed one after another, reducing
chances of missing an alarm.