Fix AlarmActivity behaviour
Previously when an alarm is dismissed from the notification,
AlarmActivity does not go away. At this time clicking "Dismiss" or
"Snooze" button in the Activity will cause a NullPointerException
because the alarm is already gone.

This commit makes sure that the AlarmActivity goes away at the same time
as the user dismiss the alarm from the notification.