Fix alarm calculation when alarms fire
Because of the nature of AlarmManager API, it is possible that some
alarms fire several seconds before the time set for it. As a result,
the next time of alarm calculated by AlarmCalc is actually this time.

To fix this problem, the time set for the alarm is passed as a parameter
when calculating the next time, ensuring that AlarmCalc returns the
correct next time.