Change reference to alarm back to position
Referencing to alarm object in the list directly causes mismatch of data
and view after switching the alarm off and on, and therefore causes a
unnecessary update to the list, makes the updated alarm blink slightly.

Move back to reference by the indices in list, but use layoutPosition
this time, so it always matches what the user selects.

Since the context menu will be dismissed automatically when not safe, it
is safe to reference by indices here.